:sparkles: feat: 본인 일정만 수정, 삭제 가능하도록 변경
세션으로 유저 정보를 가져와서 적용
- 일정의 유저 정보와 일치하는지 확인
- 본인 일정이 아니면 401
- 수정 시, 작성자 이름 미입력으로 변경

리펙토링
- 상수 LOGIN_USER 생성 및 적용
- login, logout 로직 UserService로 이동

+++ src/main/java/com/example/scheduleappdev/service/ScheduleService.java
@@ -1,11 +1,15 @@
 package com.example.scheduleappdev.service;
 
+import com.example.scheduleappdev.common.Const;
 import com.example.scheduleappdev.dto.ScheduleResDto;
+import com.example.scheduleappdev.dto.UserResDto;
 import com.example.scheduleappdev.entity.Schedule;
 import com.example.scheduleappdev.entity.User;
 import com.example.scheduleappdev.exception.UnauthorizedException;
 import com.example.scheduleappdev.repository.ScheduleRepository;
 import com.example.scheduleappdev.repository.UserRepository;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -43,21 +47,31 @@ public class ScheduleService {
     }
 
     @Transactional
-    public ScheduleResDto updateSchedule(Long id, String authorName, String todoTitle, String todoContents) {
+    public ScheduleResDto updateSchedule(Long id, String todoTitle, String todoContents, HttpServletRequest req) {
         Schedule findSchedule = scheduleRepository.findByIdOrElseThrow(id);
-        // 이름으로 비교
-        if (!findSchedule.getUser().getUserName().equals(authorName)) {
-            throw new UnauthorizedException("작성자가 일치하지 않습니다.");
-        }
+        checkScheduleAuthor(findSchedule, req);
+
         findSchedule.updateSchedule(todoTitle, todoContents);
         log.info("일정 수정 : id = {}", id);
         return new ScheduleResDto(findSchedule);
     }
 
-    public void deleteScheduleById(Long id) {
+    public void deleteScheduleById(Long id, HttpServletRequest req) {
         Schedule findSchedule = scheduleRepository.findByIdOrElseThrow(id);
+        checkScheduleAuthor(findSchedule, req);
+
         log.info("일정 삭제 : id = {}", id);
         scheduleRepository.delete(findSchedule);
+    }
+
+    public void checkScheduleAuthor(Schedule schedule, HttpServletRequest req) {
+        HttpSession session = req.getSession(false);
+        if (session != null) {
+            UserResDto loginUser = (UserResDto) session.getAttribute(Const.LOGIN_USER);
+            if (!schedule.getUser().getUserName().equals(loginUser.getUserName())) {
+                throw new UnauthorizedException("작성자가 일치하지 않습니다.");
+            }
+        }
     }
 
 }
